Move helper functions into utils.filemanip.
